### 2024-11-19 #7494 @jbytecode
smillerc - Hi, I am the handling editor of this submission. Thank you for sending your work to JOSS.

I will try to find at least 2 suitable reviewers. The editorialbot reports an AI based recommendations of similar papers in here

https://github.com/openjournals/joss-reviews/issues/7494#issuecomment-2484228206

We can select the authors of those similar submissions as reviewers. Another source is the [list of potential reviewers](https://reviewers.joss.theoj.org/lookup).

First, I will try inviting the reviewers you suggest, if any. Otherwise, I will search them myself.

Please if you have suggestions, provide me a list with their GitHub handles and please don't use `@` character to avoid any bothering notifications.

Thank you in advance!
