### 2024-11-19 #7494 @smillerc
@jbytecode Thank you for taking on the editor role! I don't really have a good idea on who to recommend, other than perhaps `hemanthgrylls`. I cite his work in the manuscript, so he will be familiar with the type of problem I'm solving with the software.
