### 2024-11-19 #7494 @jbytecode
👋👋👋 Dear @hemanthgrylls, @DanielVandH 👋👋👋

**[CurvilinearGrids.jl: A Julia package for curvilinear coordinate transformations](https://github.com/openjournals/joss-reviews/issues/7494)**

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/7494).

Thank you in advance!
