### 2024-11-19 #7494 @DanielVandH
Hi @jbytecode, sure I'm happy to.
