### 2024-11-19 #7508 @editorialbot
[REVIEW]: CurvilinearGrids.jl: A Julia package for curvilinear coordinate transformations
**Submitting author:** @smillerc (<a href="http://orcid.org/0000-0001-6452-2308">Samuel Miller</a>)
**Repository:** https://github.com/smillerc/CurvilinearGrids.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.7.1
