### 2024-11-27 #7508 @hemanthgrylls
The library is a well-structured and valuable resource for the CFD community, particularly for those developing fluid flow solvers in Julia. It is modular and facilitates integration into new solvers. The documentation is clear and user-friendly. However, there are a few points that need attention:

1. The library currently uses a non-conservative formulation for metric terms, which may violate the Geometric Conservation Law (GCL). A conservative approach, as outlined in Eq. (8) of Wang et al., 2014 or Eq. (13) of [Zhong et al., 2019](https://doi.org/10.1016/j.jcp.2019.108907), would ensure strict conservation and improve the robustness of the solver.
2. It will be useful if the authors could also add a metric term computation scheme alongside MEG6, that employs a standard fourth/sixth order central differencing scheme, which is pretty easy to implement. This can be useful for users who want to use a central differencing scheme to solve their fluid flow equations.
3. The write-up should clearly articulate the motivation for the library and the gap it fills in the CFD community, as required by JOSS guidelines.

Addressing these points, particularly the implementation of conservative metrics, would significantly enhance the library’s utility and its suitability for publication in JOSS.
