### 2024-11-27 smillerc/CurvilinearGrids.jl#24 @DanielVandH
[JOSS] Review comments
Hi @smillerc,

This is for https://github.com/openjournals/joss-reviews/issues/7508.

The paper is nicely written and your code has a nice structure, and gives a nice contribution for the Julia ecosystem. I just have some comments:

- The README should explain how to install the package (essentially, just something like `using Pkg; Pkg.add("CurvilinearGrids"); using CurvilinearGrids`).
- You use the terminology "halo" in the README and in the code example in the paper, but that's not clearly defined anywhere.
- The paper is missing a statement of need. At present, it is not immediately obvious to the reader (1) why this work is useful and (2) what gap it fills compared to other software in this area. It would improve the work to know how it compares to this other software.
- While I can roughly understand how to make use of the package from the README examples, I think it might help to also include an example where you are actually applying the package to another applied problem, which would also really showcase how the API can be used.
- The API describes `coord`, `centroid`, `coords`, and `centroids`, but you should also document the `CurvilinearGrid` functions.
- Your CI is only testing 1.9 and nightly. You should change it to test `1`, `lts`, and `pre` (or, at least, `1` so that it tests the latest release). I confirmed locally that it passes on 1.11 but it should be included in the CI.
- The definition of the Jacobian is given at the very end but maybe it'd be useful to explain it more concisely sooner since you make use of it already.

Some other minor things:

- Line 22: "Figure Figure 1"
