### 2024-11-27 df85468
update julia versions in CI
