### 2024-12-02 d1ac23f
Added better docstring for grid constructors; fixed unit test failures
