+++ README.md
@@ -7,9 +7,15 @@
 
 ![Alt text](docs/image.png)
 
-`CurvilinearGrids.jl` currently defines the `CurvilinearGrid2D` and `CurvilinearGrid3D` types. To construct these, you need to provide the discrete coordinates for the mesh. 
+`CurvilinearGrids.jl` defines `CurvilinearGrid1D`, `CurvilinearGrid2D`, and `CurvilinearGrid3D` types. To construct these, you need to provide the discrete coordinates for the mesh. 
 
-### Example
+## Installation
+`CurvilinearGrids` is a registered Julia package, so installation follows the typical procedure:
+```julia
+using Pkg; Pkg.add("CurvilinearGrids"); using CurvilinearGrids
+```
+
+### Example Grid Construction
 ```julia
 using CurvilinearGrids
 
@@ -52,7 +58,7 @@ mesh = CurvilinearGrid3D(x, y, z, nhalo)
 
 The API is still a work-in-progress, but for the moment, these functions are exported:
 
-Here `idx` can be a `Tuple` or `CartesianIndex`, and mesh is an `AbstractCurvilinearGrid`. **Important:** The indices provided to these functions are aware of halo regions, so the functions do the offsets for you. This is by design, since fields attached to the mesh, like density or pressure for example, _will_ have halo regions, and loops through these fields typically have pre-defined limits that constrain the loop to only work in the non-halo cells. If you don't use halo cells, just set `nhalo=0` in the constructors.
+Here `idx` can be a `Tuple` or `CartesianIndex`, and mesh is an `AbstractCurvilinearGrid`. **Important:** The indices provided to these functions are aware of halo regions, so the functions do the offsets for you. Halo cells, or halo regions, are an additional layer of cells around the boundary of the mesh that do not contain geometric information. Halo cells are used to apply boundary conditions in a simple manner and to exchange data between domains (in a domain-decomposed setting, e.g. using MPI). This is by design, since fields attached to the mesh, like density or pressure for example, _will_ have halo regions, and loops through these fields typically have pre-defined limits that constrain the loop to only work in the non-halo cells. If you don't use halo cells, just set `nhalo=0` in the constructors.
 
 The index can be a `Tuple`, scalar `Integer`, or `CartesianIndex`.
 - `coord(mesh, idx)`: Get the $(x,y,z)$ coordinates at index `idx`. This can be 1, 2, or 3D.
@@ -61,6 +67,24 @@ The index can be a `Tuple`, scalar `Integer`, or `CartesianIndex`.
 These functions are primarily used to get the complete set of coordinates for plotting or post-processing. These do _not_ use halo regions, since there geometry is ill-defined here.
 - `coords(mesh)` Get the: array of coordinates for the entire mesh (typically for writing to a .vtk for example)
 - `centroids(mesh)` Get the: array of centroid coordinates for the entire mesh (typically for writing to a `.vtk` file)
+
+## Constructors
+
+General purpuse constructors for 1D/2D/3D grids. These need a vector/matrix/array of vertex coordinates and the number of halo cells to pad by.
+- `CurvilinearGrid1D(x::AbstractVector{T}, nhalo::Int)`
[…]
