+++ paper/paper.md
@@ -22,7 +22,7 @@ bibliography: references.bib
 
 `CurvilinearGrids.jl` is a Julia package for computing grid metrics required for curvilinear coordinate transformations in arbitrary 1D, 2D, and 3D geometry. Curvilinear coordinate transformations are often used for finite-difference discretizations of partial differential equations, see for example, the Euler and Navier-Stokes equations in References [@visbal2002; @chandravamsi2023]. These coordinate transformations need to adhere to strict conservation laws that make the computation of grid metrics more difficult that simple finite-difference methods.
 
-The API in `CurvilinearGrids.jl`  is designed to handle the complicated book-keeping associated with metric computation on multidimensional meshes. The aim is to simplify the process by which arbitrary curvilinear geometry is incorporated into various PDE-oriented codes discretized through finite-differences. Users provide `CurvilinearGrids.jl` with discrete coordinate points defined in real-space $(x,y,z)$, or functional forms, e.g `x(i,j), y(i,j)`. These are used to compute the Jacobian matrices required to transform to a uniform computational coordinate space $(\xi,\eta,\zeta)$. A common example of this is to use a body-fit or conformal grid and transform it so that it becomes a uniform grid in $(\xi,\eta,\zeta)$. Standard finite-difference stencils can be used on the uniform transformed grid. Figure \autoref{fig:grid} shows an example of a curvilinear grid defined in real-space and it's representation in computational space. 
+The API in `CurvilinearGrids.jl`  is designed to handle the complicated book-keeping associated with metric computation on multidimensional meshes. The aim is to simplify the process by which arbitrary curvilinear geometry is incorporated into various PDE-oriented codes discretized through finite-differences. Users provide `CurvilinearGrids.jl` with discrete coordinate points defined in real-space $(x,y,z)$, or functional forms, e.g `x(i,j), y(i,j)`. These are used to compute the Jacobian matrices required to transform to a uniform computational coordinate space $(\xi,\eta,\zeta)$. A common example of this is to use a body-fit or conformal grid and transform it so that it becomes a uniform grid in $(\xi,\eta,\zeta)$. Standard finite-difference stencils can be used on the uniform transformed grid. \autoref{fig:grid} shows an example of a curvilinear grid defined in real-space and it's representation in computational space. 
 
 ![Curvilinear grid transformation.\label{fig:grid}](mesh.png)
 
