### 2024-12-02 #7508 @smillerc
@hemanthgrylls Thank you for your thorough review!

1. Thank you for letting me know about the symmetric conservative form. Can you point me to the Wang et al. reference? I was unable to find it. I was unaware of this particular form, but still use the conservative form from Visbal et. al -- note the inner/outer derivative terms (see [MonotoneExplicit6thOrderScheme.jl#L166](https://github.com/smillerc/CurvilinearGrids.jl/blob/890620184af4696d91b5220923bd48c380a44e79/src/metric_schemes/meg6/MonotoneExplicit6thOrderScheme.jl#L166)). The test at [tests/unit/test_3d.jl#L140](https://github.com/smillerc/CurvilinearGrids.jl/blob/890620184af4696d91b5220923bd48c380a44e79/test/unit/test_3d.jl#L140) was written to double-check the GCL conservation.  The edge metrics are constructed in this conservative manner, which is why they are stored at $\hat \xi$ in `edge_metrics.i₊½.ξ̂`, whereas the cell-centered metrics are _not_ conservative, e.g. (`cell_center_metrics.ξ`), and don't have the hat (just $\xi$). Maybe this could be clearer in the documentation. It would be useful to add the symmetric form as well in future releases. Is there something else I'm missing?

2. I agree that adding additional metric schemes will be helpful. This was planned for later releases -- I'd like to be able to compare different schemes. There is a partially implemented version of this in [metric_schemes/finitediff_metrics.jl](https://github.com/smillerc/CurvilinearGrids.jl/blob/main/src/metric_schemes/finitediff_metrics.jl) that needs to be incorporated into the rest of the library.

3. I'm working on updating the manuscript to include this (the other reviewer mentioned this as well).
