### 2024-12-04 b914ce2
paper / reference updates
