+++ paper/paper.md
@@ -20,9 +20,14 @@ bibliography: references.bib
 
 # Summary
 
-`CurvilinearGrids.jl` is a Julia package for computing grid metrics required for curvilinear coordinate transformations in arbitrary 1D, 2D, and 3D geometry. Curvilinear coordinate transformations are often used for finite-difference discretizations of partial differential equations, see for example, the Euler and Navier-Stokes equations in References [@visbal2002; @chandravamsi2023]. These coordinate transformations need to adhere to strict conservation laws that make the computation of grid metrics more difficult that simple finite-difference methods.
+Finite-difference discretizations of partial differential equations are widespread throughout the scientific community. Oftentimes finite-differences are used to compute spatial gradients of fields on a discrete grid, which is typically a uniform or rectilinear Cartesian mesh. *Arbitrary* multidimensional geometry is difficult to discretize directly with finite differences, however, due to non-uniform grid spacing and non-orthogonality. Curvilinear coordinate transformations can be used as a strategy to enable arbitrary geometry -- see for example, applications to the Euler and Navier-Stokes equations in References [@visbal2002; @chandravamsi2023]. While these curvilinear transformations are straightforward, the governing PDEs require additional terms (metrics) and must adhere to strict conservation laws; these criteria complicate the application of the transformation and require careful implementation. 
 
-The API in `CurvilinearGrids.jl`  is designed to handle the complicated book-keeping associated with metric computation on multidimensional meshes. The aim is to simplify the process by which arbitrary curvilinear geometry is incorporated into various PDE-oriented codes discretized through finite-differences. Users provide `CurvilinearGrids.jl` with discrete coordinate points defined in real-space $(x,y,z)$, or functional forms, e.g `x(i,j), y(i,j)`. These are used to compute the Jacobian matrices required to transform to a uniform computational coordinate space $(\xi,\eta,\zeta)$. A common example of this is to use a body-fit or conformal grid and transform it so that it becomes a uniform grid in $(\xi,\eta,\zeta)$. Standard finite-difference stencils can be used on the uniform transformed grid. \autoref{fig:grid} shows an example of a curvilinear grid defined in real-space and it's representation in computational space. 
+# Statement of Need
+
+`CurvilinearGrids.jl` is a Julia package designed to facilitate curvilinear coordinate transformations and the generation of metrics for arbitrary 1D, 2D, and 3D geometry. The API is designed for simple grid construction with minimal user effort; the backend handles the complicated book-keeping associated with metric computation and ensures strict adherance to geometric conservation laws when required. While Julia packages exist to facilitate standard finite-differences, there are few, if any, that provide conservative metrics for curvilinear transformations.
+
+
[…]
