+++ paper/references.bib
@@ -76,3 +76,18 @@
   urldate      = {2024-01-16},
   langid       = {english}
 }
+
+@book{huang2011,
+	title        = {Adaptive {{Moving Mesh Methods}}},
+	author       = {Huang, Weizhang and Russell, Robert D.},
+	year         = 2011,
+	location     = {New York, NY},
+	publisher    = {Springer},
+	series       = {Applied {{Mathematical Sciences}}},
+	volume       = 174,
+	doi          = {10.1007/978-1-4419-7916-2},
+	isbn         = {978-1-4419-7915-5 978-1-4419-7916-2},
+	urldate      = {2023-12-06},
+	langid       = {english},
+	file         = {/home/smil/Zotero/storage/N4QY696T/Huang and Russell - 2011 - Adaptive Moving Mesh Methods.pdf}
+}
