### 2024-12-04 5082306
add funding section
