### 2024-12-04 smillerc/CurvilinearGrids.jl#24 @smillerc
@DanielVandH I've made changes to the paper as well as the README and internal docstrings. Thank you for your edits!
