### 2024-12-04 smillerc/CurvilinearGrids.jl#24 @DanielVandH
Thanks @smillerc, I'll take another look soon!
