### 2024-12-08 #7508 @hemanthgrylls
Hi @smillerc,
1. Sorry I did not hyperlink the reference (Its actually Nonomura et al., not Wang et al.,  sorry). This is the paper that I was referring to:
https://www.sciencedirect.com/science/article/pii/S0045793014003624

2. The hat metrics at cell faces represent Jacobian normalized to my knowledge. Please cross check with https://www.sciencedirect.com/science/article/pii/S0045793023000841

I think including this symmetric form will make it more comprehensive. But overall, It is a nice repository and useful for Julia CFD developers.

Thanks
