+++ src/grids/GridTypes.jl
@@ -77,15 +77,29 @@ include("2d.jl")
 include("3d.jl")
 include("simple_constructors/simple_constructors.jl")
 
-"""Get the size of the grid for cell-based arrays"""
-cellsize(mesh::AbstractCurvilinearGrid1D) = (mesh.nnodes - 1,)
+"""
+    cellsize(mesh::AbstractCurvilinearGrid)
+
+The total number of cells in the mesh (_excluding_ halo regions)
+"""
 cellsize(mesh::AbstractCurvilinearGrid) = @. mesh.nnodes - 1
+cellsize(mesh::AbstractCurvilinearGrid1D) = (mesh.nnodes - 1,)
 
-"""Get the size of the grid for cell-based arrays when the halo cells are included"""
-cellsize_withhalo(mesh::AbstractCurvilinearGrid1D) = (mesh.nnodes - 1 + 2 * mesh.nhalo,)
+"""
+    cellsize_withhalo(mesh::AbstractCurvilinearGrid)
+
+The total number of cells in the mesh (_including_ halo regions)
+"""
 cellsize_withhalo(mesh::AbstractCurvilinearGrid) = @. mesh.nnodes - 1 + 2 * mesh.nhalo
+cellsize_withhalo(mesh::AbstractCurvilinearGrid1D) = (mesh.nnodes - 1 + 2 * mesh.nhalo,)
 
-@inline function coords(mesh::AbstractCurvilinearGrid1D)
+"""
+    coords(mesh)
+
+Return a view of the mesh coordinates (does not include halo regions). Returns an NTuple, e.g. (x, y, z) 
+where N is the dimension of the mesh.
+"""
+function coords(mesh::AbstractCurvilinearGrid1D)
   return @views mesh.node_coordinates.x[mesh.iterators.node.domain]
 end
 
@@ -104,6 +118,12 @@ end
   )
 end
 
+"""
+    centroids(mesh)
+
+Return a view of the cell-centroids of the mesh (does not include halo regions). Returns an NTuple, e.g. (x, y, z) 
+where N is the dimension of the mesh.
+"""
 @inline function centroids(mesh::AbstractCurvilinearGrid1D)
   return @views mesh.centroid_coordinates.x[mesh.iterators.cell.domain]
 end
@@ -124,6 +144,8 @@ end
 end
 
 """
+    coord(mesh, index::CartesianIndex)
+
 Get the position/coordinate at a given index. **NOTE:** these indices are consistent with halo cells included.
 This means that if your grid has 2 halo cells, the position of the first non-halo vertex is
 index at coord(mesh, 3). The `CurvilinearGrid` only keeps track of the number of halo cells for each
@@ -131,16 +153,29 @@ dimension, whereas the grid functions have no knowledge halos. Therefore, the `c
 applies a shift to the index for you.
 """
 @inline coord(mesh, CI::CartesianIndex) = coord(mesh, CI.I)
+
+"""
+    coord(mesh, i::Int)
+"""
 @inline coord(mesh, i::Int) = coord(mesh, (i,))
 
-@inline function coord(mesh, (i,)::NTuple{1,Int})
+"""
+    coord(mesh, (i,)::NTuple{1,Int})
+"""
+function coord(mesh, (i,)::NTuple{1,Int})
   @SVector [mesh.node_coordinates.x[i]]
 end
 
+"""
+    coord(mesh, (i, j)::NTuple{2,Int})
+"""
 @inline function coord(mesh, (i, j)::NTuple{2,Int})
   @SVector [mesh.node_coordinates.x[i, j], mesh.node_coordinates.y[i, j]]
[…]
