### 2024-12-12 ee20ddf
refactor to new metric and discretization org/scheme
