+++ src/discretization_schemes/metrics.jl
@@ -1,0 +1,476 @@
+
+using MappedArrays
+using UnPack
+
+"""
+    forward_metrics!(scheme::DiscretizationScheme, metrics, x::AbstractArray{T,1}) where {T}
+
+Compute the forward metrics ∂x/∂ξ, or `x_ξ` based on the centroid coordinates `x`.
+"""
+function forward_metrics!(
+  scheme::DiscretizationScheme, metrics, x::AbstractArray{T,1}
+) where {T}
+  # axes
+  ξ = 1
+  domain = CartesianIndices(x)
+
+  @views begin
+    @unpack x_ξ = metrics
+  end
+
+  cell_center_derivatives!(scheme, x_ξ, x, ξ, domain)
+
+  return nothing
+end
+
+"""
+    forward_metrics!(scheme::DiscretizationScheme, metrics, x::AbstractArray{T,2}, y::AbstractArray{T,2}) where {T}
+
+Compute the forward metrics ∂xᵢ/∂ξᵢ, i.e., `x_ξ, x_η, y_ξ, y_η`, based on the centroid coordinates `x` and `y`.
+"""
+function forward_metrics!(
+  scheme::DiscretizationScheme,
+  metrics,
+  x::AbstractArray{T,2},
+  y::AbstractArray{T,2},
+  domain,
+) where {T}
+
+  # axes
+  ξ, η = (1, 2)
+
+  if size(x) != size(y)
+    error("Size mismatch for the given x, y coordinate arrays, they must all be the same")
+  end
+
+  @views begin
+    xξ = metrics.x₁.ξ
+    xη = metrics.x₁.η
+    yξ = metrics.x₂.ξ
+    yη = metrics.x₂.η
+  end
+
+  cell_center_derivatives!(scheme, xξ, x, ξ, domain)
+  cell_center_derivatives!(scheme, xη, x, η, domain)
+  cell_center_derivatives!(scheme, yξ, y, ξ, domain)
+  cell_center_derivatives!(scheme, yη, y, η, domain)
+
+  @kernel inbounds = true function jacobian_2d_kernel!(x_ξ, x_η, y_ξ, y_η, jacobian, itr)
+    I = @index(Global, Cartesian)
+    idx = itr[I]
+    _jacobian_matrix = @SMatrix [
+      x_ξ[idx] x_η[idx]
+      y_ξ[idx] y_η[idx]
+    ]
+
+    jacobian[idx] = det(_jacobian_matrix)
+  end
+
+  jacobian_2d_kernel!(scheme.backend)(
+    xξ, xη, yξ, yη, metrics.J, domain; ndrange=size(domain)
+  )
+
+  return nothing
+end
+
+"""
+    forward_metrics!(scheme::DiscretizationScheme, metrics, x::AbstractArray{T,3}, y::AbstractArray{T,3}, z::AbstractArray{T,3}) where {T}
+
+Compute the forward metrics ∂xᵢ/∂ξᵢ, i.e., `x_ξ, x_η, x_ζ, y_ξ, y_η, y_ζ, z_ξ, z_η, z_ζ`, based on the centroid coordinates `x` and `y`, and `z`.
+
+"""
+function forward_metrics!(
+  scheme::DiscretizationScheme,
+  metrics,
+  x::AbstractArray{T,3},
+  y::AbstractArray{T,3},
+  z::AbstractArray{T,3},
+) where {T}
+
+  # axes
[…]
