+++ src/discretization_schemes/meg/MonotoneExplicitGradients.jl
@@ -1,0 +1,333 @@
+module MontoneExplicitGradientSchemes
+
+using StaticArrays
+using KernelAbstractions
+using MappedArrays
+using CartesianDomains
+using MappedArrays
+
+using ..DiscretizationSchemes
+using ..DiscretizationSchemes: DiscretizationScheme
+
+export MontoneExplicitGradientScheme6thOrder
+export reconstruct_interface,
+  interpolate_to_edge,
+  interface_derivative,
+  interface_derivative_nodamping,
+  cell_center_derivative
+
+export cell_center_derivatives!, interpolate_to_edge!
+
+struct MontoneExplicitGradientScheme6thOrder{C,B} <: DiscretizationScheme
+  cache::C
+  backend::B
+  nhalo::Int
+end
+
+# The MEG6 scheme requires a halo of 5 cells in all dimensions
+
+function MontoneExplicitGradientScheme6thOrder(;
+  use_cache=true, celldims=nothing, backend=CPU(), T=Float64
+)
+  nhalo = 5
+  if use_cache && isnothing(celldims)
+    error(
+      "When use_cache=true (default), celldims must be an NTuple providing the dimensions of the cell-based domain",
+    )
+  end
+
+  if use_cache
+    cache = (;
+      ∂²ϕ=KernelAbstractions.zeros(backend, T, celldims),
+      ∂ϕ=KernelAbstractions.zeros(backend, T, celldims),
+      outer_deriv_1=KernelAbstractions.zeros(backend, T, celldims),
+      outer_deriv_2=KernelAbstractions.zeros(backend, T, celldims),
+    )
+  else
+    cache = nothing
+  end
+
+  return MontoneExplicitGradientScheme6thOrder(cache, backend, nhalo)
+end
+
+include("cell_center_derivs.jl")
+include("interp_to_edges.jl")
+
+reconstruct_interface(::MontoneExplicitGradientScheme6thOrder, ϕ...) = meg6_ϕᴸᴿᵢ₊½(ϕ...)
+
+interpolate_to_edge(::MontoneExplicitGradientScheme6thOrder, ϕ...) = meg6_ϕᵢ₊½(ϕ...)
+
[…]
