+++ test/unit/test_2d.jl
@@ -1,39 +1,46 @@
-@testset "2D Rectangular Mesh" begin
+# @testset "2D Rectangular Mesh" 
+begin
   include("common.jl")
 
-  ni, nj = (4, 8)
-  nhalo = 2
+  ni, nj = (40, 80)
+  nhalo = 5
   x0, x1 = (0, 2)
   y0, y1 = (1, 3)
   mesh = RectlinearGrid((x0, y0), (x1, y1), (ni, nj), nhalo)
   domain = mesh.iterators.cell.domain
 
-  @test mesh.iterators.cell.full == CartesianIndices((8, 12))
-  @test mesh.iterators.cell.domain == CartesianIndices((3:6, 3:10))
+  display(mesh.centroid_coordinates.x)
 
-  @test mesh.iterators.node.full == CartesianIndices((9, 13))
-  @test mesh.iterators.node.domain == CartesianIndices((3:7, 3:11))
+  println()
+  display(mesh.cell_center_metrics.forward.x₁.ξ)
+  # display(mesh.cell_center_metrics.forward.J[domain])
 
-  @test mesh.domain_limits.node == (ilo=3, ihi=7, jlo=3, jhi=11)
-  @test mesh.domain_limits.cell == (ilo=3, ihi=6, jlo=3, jhi=10)
+  # @test mesh.iterators.cell.full == CartesianIndices((8, 12))
+  # @test mesh.iterators.cell.domain == CartesianIndices((3:6, 3:10))
 
-  @test all(mesh.cell_center_metrics.J[domain] .≈ 0.125)
-  @test all(mesh.cell_center_metrics.x₁.ξ[domain] .≈ 0.5)
-  @test all(mesh.cell_center_metrics.x₂.ξ[domain] .≈ 0.0)
-  @test all(mesh.cell_center_metrics.x₁.η[domain] .≈ 0.0)
-  @test all(mesh.cell_center_metrics.x₂.η[domain] .≈ 0.25)
-  @test all(mesh.cell_center_metrics.ξ.x₁[domain] .≈ 2.0)
-  @test all(mesh.cell_center_metrics.ξ.x₂[domain] .≈ -0.0)
-  @test all(mesh.cell_center_metrics.η.x₁[domain] .≈ -0.0)
-  @test all(mesh.cell_center_metrics.η.x₂[domain] .≈ 4.0)
-  @test all(mesh.cell_center_metrics.ξ.t[domain] .≈ 0.0)
-  @test all(mesh.cell_center_metrics.η.t[domain] .≈ 0.0)
+  # @test mesh.iterators.node.full == CartesianIndices((9, 13))
+  # @test mesh.iterators.node.domain == CartesianIndices((3:7, 3:11))
 
-  @test all(mesh.edge_metrics.i₊½.J[domain] .≈ 0.125)
-  @test all(mesh.edge_metrics.i₊½.ξ̂.x₁[domain] .≈ 0.25)
-  @test all(mesh.edge_metrics.i₊½.ξ̂.x₂[domain] .≈ 0.0)
-  @test all(mesh.edge_metrics.i₊½.η̂.x₁[domain] .≈ 0.0)
-  @test all(mesh.edge_metrics.i₊½.η̂.x₂[domain] .≈ 0.5)
+  # @test mesh.domain_limits.node == (ilo=3, ihi=7, jlo=3, jhi=11)
+  # @test mesh.domain_limits.cell == (ilo=3, ihi=6, jlo=3, jhi=10)
+
[…]
