### 2024-12-13 ec46edd
more refactoring with updated metric scheme; 2d tests are passing now
