+++ src/grids/metric_soa.jl
@@ -64,72 +64,59 @@ function get_metric_soa(celldims::NTuple{3,Int}, backend, T)
   edge_metrics = (
     inverse=(
       i₊½=(
-        J=KernelAbstractions.zeros(backend, T, celldims),
         ξ=StructArray((
           x₁=KernelAbstractions.zeros(backend, T, celldims),
           x₂=KernelAbstractions.zeros(backend, T, celldims),
           x₃=KernelAbstractions.zeros(backend, T, celldims),
-          t=KernelAbstractions.zeros(backend, T, celldims),
         )),
         η=StructArray((
           x₁=KernelAbstractions.zeros(backend, T, celldims),
           x₂=KernelAbstractions.zeros(backend, T, celldims),
           x₃=KernelAbstractions.zeros(backend, T, celldims),
-          t=KernelAbstractions.zeros(backend, T, celldims),
         )),
         ζ=StructArray((
           x₁=KernelAbstractions.zeros(backend, T, celldims),
           x₂=KernelAbstractions.zeros(backend, T, celldims),
           x₃=KernelAbstractions.zeros(backend, T, celldims),
-          t=KernelAbstractions.zeros(backend, T, celldims),
         )),
       ),
       j₊½=(
-        J=KernelAbstractions.zeros(backend, T, celldims),
         ξ=StructArray((
           x₁=KernelAbstractions.zeros(backend, T, celldims),
           x₂=KernelAbstractions.zeros(backend, T, celldims),
           x₃=KernelAbstractions.zeros(backend, T, celldims),
-          t=KernelAbstractions.zeros(backend, T, celldims),
         )),
         η=StructArray((
           x₁=KernelAbstractions.zeros(backend, T, celldims),
           x₂=KernelAbstractions.zeros(backend, T, celldims),
           x₃=KernelAbstractions.zeros(backend, T, celldims),
-          t=KernelAbstractions.zeros(backend, T, celldims),
         )),
         ζ=StructArray((
           x₁=KernelAbstractions.zeros(backend, T, celldims),
           x₂=KernelAbstractions.zeros(backend, T, celldims),
           x₃=KernelAbstractions.zeros(backend, T, celldims),
-          t=KernelAbstractions.zeros(backend, T, celldims),
         )),
       ),
       k₊½=(
-        J=KernelAbstractions.zeros(backend, T, celldims),
         ξ=StructArray((
           x₁=KernelAbstractions.zeros(backend, T, celldims),
           x₂=KernelAbstractions.zeros(backend, T, celldims),
           x₃=KernelAbstractions.zeros(backend, T, celldims),
-          t=KernelAbstractions.zeros(backend, T, celldims),
         )),
         η=StructArray((
           x₁=KernelAbstractions.zeros(backend, T, celldims),
           x₂=KernelAbstractions.zeros(backend, T, celldims),
           x₃=KernelAbstractions.zeros(backend, T, celldims),
-          t=KernelAbstractions.zeros(backend, T, celldims),
         )),
         ζ=StructArray((
           x₁=KernelAbstractions.zeros(backend, T, celldims),
           x₂=KernelAbstractions.zeros(backend, T, celldims),
           x₃=KernelAbstractions.zeros(backend, T, celldims),
-          t=KernelAbstractions.zeros(backend, T, celldims),
         )),
       ),
     ),
     inverse_normalized=(
       i₊½=(
-        J=KernelAbstractions.zeros(backend, T, celldims),
         ξ̂=StructArray((
           x₁=KernelAbstractions.zeros(backend, T, celldims),
[…]
