+++ src/grids/2d.jl
@@ -354,14 +354,12 @@ end
 
 function update_metrics!(mesh::AbstractCurvilinearGrid2D, t::Real=0)
   # Update the metrics within the non-halo region, e.g., the domain
-  domain = mesh.iterators.cell.domain
-
   MetricDiscretizationSchemes.update_metrics!(
     mesh.discretization_scheme,
     mesh.centroid_coordinates,
     mesh.cell_center_metrics,
     mesh.edge_metrics,
-    domain,
+    mesh.iterators.cell.domain,
   )
 
   return nothing
