### 2024-12-13 4cbd746
update the `to_vtk` function for 3D
