### 2024-12-13 d94afbb
fix the API so `nhalo` doesn't get set by the user, but rather by the discretization scheme the user picks -- this is less error prone
