+++ src/grids/simple_constructors/rectlinear.jl
@@ -1,13 +1,13 @@
 
 """
-    RectlinearGrid((x0, x1), ncells, nhalo; backend=CPU(), T=Float64, is_static=true, make_uniform=false, tile_layout=nothing, rank::Int=-1)
+    RectlinearGrid((x0, x1), ncells, discretization_scheme, backend=CPU(), T=Float64, is_static=true, make_uniform=false, tile_layout=nothing, rank::Int=-1)
 
 Generate a 1D rectlinear grid using start/end points and cell resolution.
 """
 function RectlinearGrid(
   (x0, x1),
   ncells,
-  nhalo::Int;
+  discretization_scheme::Symbol;
   backend=CPU(),
   T=Float64,
   is_static=true,
@@ -17,17 +17,17 @@ function RectlinearGrid(
 )
   ni = ncells + 1
   x = collect(T, range(x0, x1; length=ni))
-  return CurvilinearGrid1D(x, nhalo; backend=backend)
+  return CurvilinearGrid1D(x, discretization_scheme; backend=backend)
 end
 
 """
-    RectlinearGrid(x, nhalo; backend=CPU(), is_static=true, make_uniform=false, tile_layout=nothing, rank::Int=-1)
+    RectlinearGrid(x, discretization_scheme, backend=CPU(), is_static=true, make_uniform=false, tile_layout=nothing, rank::Int=-1)
 
 Generate a 1D rectlinear grid using an x coordinate vector.
 """
 function RectlinearGrid(
   x::AbstractVector{T},
-  nhalo::Int;
+  discretization_scheme::Symbol;
   backend=CPU(),
   is_static=true,
   make_uniform=false,
@@ -75,7 +75,7 @@ function RectlinearGrid(
 
     return CurvilinearGrid1D(
       x_local,
-      nhalo;
+      discretization_scheme;
       backend=backend,
       on_bc=on_bc,
       tiles=tiled_node_limits,
@@ -85,7 +85,7 @@ function RectlinearGrid(
   else
     return CurvilinearGrid1D(
       x,
-      nhalo;
+      discretization_scheme;
       backend=backend,
       is_orthogonal=true,
       is_static=is_static,
@@ -95,14 +95,14 @@ function RectlinearGrid(
 end
 
 """
-    RectlinearCylindricalGrid((r0, r1), ncells, nhalo; snap_to_axis=true, backend=CPU(), T=Float64, is_static=true, make_uniform=false, tile_layout=nothing, rank::Int=-1)
+    RectlinearCylindricalGrid((r0, r1), ncells, discretization_scheme, snap_to_axis=true, backend=CPU(), T=Float64, is_static=true, make_uniform=false, tile_layout=nothing, rank::Int=-1)
 
 Generate a 1D cyclindrical rectlinear grid using start/end points and cell resolution.
 """
 function RectlinearCylindricalGrid(
   (r0, r1),
   ncells,
-  nhalo::Int;
+  discretization_scheme::Symbol;
   snap_to_axis=true,
   backend=CPU(),
   T=Float64,
@@ -113,11 +113,11 @@ function RectlinearCylindricalGrid(
 )
   ni = ncells + 1
   r = collect(T, range(r0, r1; length=ni))
-  return CylindricalGrid1D(r, nhalo, snap_to_axis; backend=backend)
+  return CylindricalGrid1D(r, discretization_scheme, snap_to_axis; backend=backend)
 end
 
 """
-    RectlinearSphericalGrid((r0, r1), ncells, nhalo; snap_to_axis=true, backend=CPU(), T=Float64.is_static = true, make_uniform=false, tile_layout=nothing, rank::Int=-1,
[…]

[…]
