+++ src/discretization_schemes/metrics.jl
@@ -55,9 +55,9 @@ function forward_metrics!(
   cell_center_derivatives!(scheme, yξ, y, ξ, domain)
   cell_center_derivatives!(scheme, yη, y, η, domain)
 
-  @kernel inbounds = true function jacobian_2d_kernel!(x_ξ, x_η, y_ξ, y_η, jacobian, itr)
+  @kernel inbounds = true function jacobian_2d_kernel!(x_ξ, x_η, y_ξ, y_η, jacobian, domain)
     I = @index(Global, Cartesian)
-    idx = itr[I]
+    idx = domain[I]
     _jacobian_matrix = @SMatrix [
       x_ξ[idx] x_η[idx]
       y_ξ[idx] y_η[idx]
@@ -85,6 +85,7 @@ function forward_metrics!(
   x::AbstractArray{T,3},
   y::AbstractArray{T,3},
   z::AbstractArray{T,3},
+  domain,
 ) where {T}
 
   # axes
@@ -96,23 +97,50 @@ function forward_metrics!(
     )
   end
 
-  domain = CartesianIndices(x)
-
   @views begin
-    @unpack x_ξ, x_η, x_ζ, y_ξ, y_η, y_ζ, z_ξ, z_η, z_ζ = metrics
+    xξ = metrics.x₁.ξ
+    xη = metrics.x₁.η
+    xζ = metrics.x₁.ζ
+
+    yξ = metrics.x₂.ξ
+    yη = metrics.x₂.η
+    yζ = metrics.x₂.ζ
+
+    zξ = metrics.x₃.ξ
+    zη = metrics.x₃.η
+    zζ = metrics.x₃.ζ
   end
 
-  cell_center_derivatives!(scheme, x_ξ, x, ξ, domain)
-  cell_center_derivatives!(scheme, x_η, x, η, domain)
-  cell_center_derivatives!(scheme, x_ζ, x, ζ, domain)
-  cell_center_derivatives!(scheme, y_ξ, y, ξ, domain)
-  cell_center_derivatives!(scheme, y_η, y, η, domain)
-  cell_center_derivatives!(scheme, y_ζ, y, ζ, domain)
-  cell_center_derivatives!(scheme, z_ξ, z, ξ, domain)
-  cell_center_derivatives!(scheme, z_η, z, η, domain)
-  cell_center_derivatives!(scheme, z_ζ, z, ζ, domain)
+  cell_center_derivatives!(scheme, xξ, x, ξ, domain)
+  cell_center_derivatives!(scheme, xη, x, η, domain)
+  cell_center_derivatives!(scheme, xζ, x, ζ, domain)
+  cell_center_derivatives!(scheme, yξ, y, ξ, domain)
+  cell_center_derivatives!(scheme, yη, y, η, domain)
+  cell_center_derivatives!(scheme, yζ, y, ζ, domain)
+  cell_center_derivatives!(scheme, zξ, z, ξ, domain)
+  cell_center_derivatives!(scheme, zη, z, η, domain)
+  cell_center_derivatives!(scheme, zζ, z, ζ, domain)
+
+  jacobian_3d_kernel!(scheme.backend)(
+    xξ, xη, xζ, yξ, yη, yζ, zξ, zη, zζ, metrics.J, domain; ndrange=size(domain)
+  )
 
   return nothing
+end
+
+# The compute kernel to find the jacobian in 3D
[…]
