### 2024-12-13 35c0f01
update the metric SoA layout
