### 2024-12-13 f58eb6e
update the `to_vtk` function for 2D and 3D
