### 2024-12-13 d48e408
forgot to add the symmetric metric changes
