### 2024-12-13 1cbd9a9
test updates (2d and 3d are working now)
