+++ test/runtests.jl
@@ -17,6 +17,6 @@ using UnPack
   # include("unit/test_2d_axisymmetric.jl")
   # include("unit/perturb_mesh.jl")
 
-  # include("unit/test_3d.jl")
+  include("unit/test_3d.jl")
   # include("unit/test_meg6.jl")
 end
