+++ test/unit/test_3d.jl
@@ -5,165 +5,131 @@
   x0, x1 = (0.0, 2.0)
   y0, y1 = (1, 3)
   z0, z1 = (-1, 2)
-  ni, nj, nk = (4, 8, 12)
-  nhalo = 4
+  ni, nj, nk = (40, 80, 120)
 
-  mesh = RectlinearGrid((x0, y0, z0), (x1, y1, z1), (ni, nj, nk), nhalo)
+  mesh = RectlinearGrid((x0, y0, z0), (x1, y1, z1), (ni, nj, nk), :meg6)
   domain = mesh.iterators.cell.domain
 
-  cell_volume = 0.5 * 0.25 * 0.25
+  cell_volume = 0.05 * 0.025 * 0.025
 
-  @test all(mesh.cell_center_metrics.J[domain] .== cell_volume)
-  @test all(mesh.cell_center_metrics.ξ.x₁[domain] .== 2.0)
-  @test all(mesh.cell_center_metrics.ξ.x₂[domain] .== 0.0)
-  @test all(mesh.cell_center_metrics.ξ.x₃[domain] .== 0.0)
-  @test all(mesh.cell_center_metrics.η.x₁[domain] .== 0.0)
-  @test all(mesh.cell_center_metrics.η.x₂[domain] .== 4.0)
-  @test all(mesh.cell_center_metrics.η.x₃[domain] .== 0.0)
-  @test all(mesh.cell_center_metrics.ζ.x₁[domain] .== 0.0)
-  @test all(mesh.cell_center_metrics.ζ.x₂[domain] .== 0.0)
-  @test all(mesh.cell_center_metrics.ζ.x₃[domain] .== 4.0)
-  @test all(mesh.cell_center_metrics.ξ.t[domain] .== 0.0)
-  @test all(mesh.cell_center_metrics.η.t[domain] .== 0.0)
-  @test all(mesh.cell_center_metrics.ζ.t[domain] .== 0.0)
+  @test all(mesh.cell_center_metrics.forward.J[domain] .≈ cell_volume)
 
-  @test all(mesh.edge_metrics.i₊½.ξ̂.x₁[domain] .≈ 0.0625)
-  @test all(mesh.edge_metrics.j₊½.ξ̂.x₁[domain] .≈ 0.0625)
-  @test all(mesh.edge_metrics.k₊½.ξ̂.x₁[domain] .≈ 0.0625)
-  @test all(mesh.edge_metrics.i₊½.ξ̂.x₂[domain] .≈ 0.0)
-  @test all(mesh.edge_metrics.j₊½.ξ̂.x₂[domain] .≈ 0.0)
-  @test all(mesh.edge_metrics.k₊½.ξ̂.x₂[domain] .≈ 0.0)
-  @test all(mesh.edge_metrics.i₊½.ξ̂.x₃[domain] .≈ 0.0)
-  @test all(mesh.edge_metrics.j₊½.ξ̂.x₃[domain] .≈ 0.0)
-  @test all(mesh.edge_metrics.k₊½.ξ̂.x₃[domain] .≈ 0.0)
-  @test all(mesh.edge_metrics.i₊½.η̂.x₁[domain] .≈ 0.0)
-  @test all(mesh.edge_metrics.j₊½.η̂.x₁[domain] .≈ 0.0)
-  @test all(mesh.edge_metrics.k₊½.η̂.x₁[domain] .≈ 0.0)
-  @test all(mesh.edge_metrics.i₊½.η̂.x₂[domain] .≈ 0.125)
[…]
