### 2024-12-13 f4a421d
refactor common method
