+++ src/grids/1d.jl
@@ -59,13 +59,22 @@ Construct a curvilinear grid in 1D using a vector of x coordinate points.
 """
 function CurvilinearGrid1D(
   x::AbstractVector{T},
-  nhalo;
+  discretization_scheme::Symbol;
   backend=CPU(),
-  discretization_scheme=:MEG6,
   on_bc=nothing,
   is_static=false,
   tiles=nothing,
 ) where {T}
+
+  #
+  if Symbol(uppercase("$discretization_scheme")) === :MEG6 ||
+    discretization_scheme == :MontoneExplicitGradientScheme6thOrder
+    MetricDiscretizationScheme = MontoneExplicitGradientScheme6thOrder
+    nhalo = 5
+  else
+    error("Only MontoneExplicitGradientScheme6thOrder or MEG6 is supported for now")
+  end
+
   ni = length(x)
   ncells = ni - 1
   ilo = nhalo + 1
@@ -90,18 +99,9 @@ function CurvilinearGrid1D(
 
   node_velocities = StructArray((x=KernelAbstractions.zeros(backend, T, nodedims),))
 
-  # if discretization_scheme === :MEG6 || discretization_scheme === :MonotoneExplicit6thOrder
-  #   if nhalo < 4
-  #     error("`nhalo` must = 4 when using the MEG6 discretization scheme")
-  #   end
-
-  discr_scheme = MetricDiscretizationSchemes.MonotoneExplicit6thOrderDiscretization(
-    domain_iterators.cell.full, T
+  discr_scheme = MetricDiscretizationScheme(;
+    use_cache=true, celldims=size(domain_iterators.cell.full), backend=backend, T=T
   )
-
-  # else
-  #   error("Unknown discretization scheme to compute the conserved metrics")
-  # end
 
   if isnothing(on_bc)
     _on_bc = (ilo=true, ihi=true)
@@ -137,14 +137,23 @@ Construct a curvilinear grid in 1D with spherical symmetry using a vector of x c
 """
 function SphericalGrid1D(
   x::AbstractVector{T},
-  nhalo::Int,
+  discretization_scheme::Symbol,
   snap_to_axis::Bool;
   backend=CPU(),
-  discretization_scheme=:MEG6,
   on_bc=nothing,
   is_static=false,
   tiles=nothing,
 ) where {T}
+
+  #
+  if Symbol(uppercase("$discretization_scheme")) === :MEG6 ||
+    discretization_scheme == :MontoneExplicitGradientScheme6thOrder
+    MetricDiscretizationScheme = MontoneExplicitGradientScheme6thOrder
+    nhalo = 5
+  else
+    error("Only MontoneExplicitGradientScheme6thOrder or MEG6 is supported for now")
+  end
+
   ni = length(x)
   ncells = ni - 1
   ilo = nhalo + 1
@@ -169,18 +178,9 @@ function SphericalGrid1D(
 
   node_velocities = StructArray((x=KernelAbstractions.zeros(backend, T, nodedims),))
 
-  # if discretization_scheme === :MEG6 || discretization_scheme === :MonotoneExplicit6thOrder
-  #   if nhalo < 4
-  #     error("`nhalo` must = 4 when using the MEG6 discretization scheme")
-  #   end
-
-  discr_scheme = MetricDiscretizationSchemes.MonotoneExplicit6thOrderDiscretization(
-    domain_iterators.cell.full, T
+  discr_scheme = MetricDiscretizationScheme(;
+    use_cache=true, celldims=size(domain_iterators.cell.full), backend=backend, T=T
   )
-
-  # else
[…]
