+++ src/grids/3d.jl
@@ -157,19 +157,6 @@ function update!(mesh::CurvilinearGrid3D; force=false)
   return nothing
 end
 
-function update_metrics!(mesh::CurvilinearGrid3D, t::Real=0)
-  # Update the metrics within the non-halo region, e.g., the domain
-  MetricDiscretizationSchemes.update_metrics!(
-    mesh.discretization_scheme,
-    mesh.centroid_coordinates,
-    mesh.cell_center_metrics,
-    mesh.edge_metrics,
-    mesh.iterators.cell.domain,
-  )
-
-  return nothing
-end
-
 # ------------------------------------------------------------------
 # Grid Metrics
 # ------------------------------------------------------------------
