+++ test/runtests.jl
@@ -8,8 +8,6 @@ using WriteVTK
 using UnPack
 
 @testset verbose = true "UnitTests" begin
-  # include("unit/test_indexing.jl")
-
   include("unit/test_1d.jl")
   # # include("unit/test_1d_axisymmetric.jl")
 
