+++ src/metric_schemes/meg6/gradients.jl
@@ -1,316 +1,0 @@
-function ∂!(
-  ∂x::AbstractArray{T,N}, x::AbstractArray{T,N}, domain, axis::Int, ϵ=5eps(T)
-) where {T,N}
-
-  # for small domains, skip, since the deriv will be zero
-  if size(domain, axis) <= 2
-    error("domain too small!!!!")
-    fill!(∂x, 0)
-    return nothing
-  end
-
-  # do edges
-  ∂_at_lobc!(∂x, x, domain, axis)
-  ∂_at_hibc!(∂x, x, domain, axis)
-
-  # coefficients; T() is used to convert to the appropriate datatype,
-  # which is more impactful on GPUs, e.g. Float32 is significantly faster.
-  # These are all compile-time constants, so the cost is nothing!
-  a = T(3 / 4)
-  b = T(3 / 20)
-  c = T(1 / 60)
-
-  if 2 < size(domain, axis) <= 4 # 2nd order
-    inner_domain = expand(domain, axis, -1)
-    # @info "2nd order"
-    @batch for i in inner_domain
-      ᵢ₋₁ = down(i, axis, 1)
-      ᵢ₊₁ = up(i, axis, 1)
-      _∂x = (x[ᵢ₊₁] - x[ᵢ₋₁]) / 2
-      ∂x[i] = _∂x * (abs(_∂x) >= ϵ)
-    end
-
-  elseif 4 < size(domain, axis) <= 7 # 4th order
-    # @info "4th order"
-    inner_domain = expand(domain, axis, -2)
-    @batch for i in inner_domain
-      ᵢ₋₂ = down(i, axis, 2)
-      ᵢ₋₁ = down(i, axis, 1)
-      ᵢ₊₁ = up(i, axis, 1)
-      ᵢ₊₂ = up(i, axis, 2)
-
-      _∂x = (-x[ᵢ₊₂] + 8x[ᵢ₊₁] - 8x[ᵢ₋₁] + x[ᵢ₋₂]) / 12
-      ∂x[i] = _∂x * (abs(_∂x) >= ϵ)
-    end
-
-  else # 6th order; size(domain, axis) > 7
-    inner_domain = expand(domain, axis, -3)
-    @batch for i in inner_domain
-      ᵢ₋₃ = down(i, axis, 3)
-      ᵢ₋₂ = down(i, axis, 2)
-      ᵢ₋₁ = down(i, axis, 1)
-
-      ᵢ₊₁ = up(i, axis, 1)
-      ᵢ₊₂ = up(i, axis, 2)
-      ᵢ₊₃ = up(i, axis, 3)
-
-      _∂x = (a * (x[ᵢ₊₁] - x[ᵢ₋₁]) - b * (x[ᵢ₊₂] - x[ᵢ₋₂]) + c * (x[ᵢ₊₃] - x[ᵢ₋₃]))
-      ∂x[i] = _∂x * (abs(_∂x) >= ϵ)
-    end
-  end
-
-  return nothing
-end
-
-function ∂²!(
-  ∂²x::AbstractArray{T,N},
-  ∂x::AbstractArray{T,N},
-  x::AbstractArray{T,N},
-  domain,
-  axis::Int,
-  ϵ=10eps(T),
-) where {T,N}
-
-  # for small domains, skip, since the deriv will be zero
-  if size(domain, axis) <= 2
-    fill!(∂²x, 0)
-    return nothing
-  end
-
-  # do edges
-  ∂²_at_lobc!(∂²x, ∂x, x, domain, axis)
-  ∂²_at_hibc!(∂²x, ∂x, x, domain, axis)
-
[…]
