### 2024-12-13 d01333b
refactor grid types
