### 2024-12-13 23254ea
throw an error if the domain is too small for meg6 to work
