### 2024-12-13 8374290
test updates; consolidate gcl check, re-enable axisym tests
