### 2024-12-16 72996b1
re-enabled metric validation checks and exported the `jacobian_matrix` function
