+++ src/grids/3d.jl
@@ -157,99 +157,23 @@ function update!(mesh::CurvilinearGrid3D; force=false)
   return nothing
 end
 
-# ------------------------------------------------------------------
-# Grid Metrics
-# ------------------------------------------------------------------
+function jacobian_matrix(mesh::CurvilinearGrid3D, (i, j, k))
+  xξ = mesh.cell_center_metrics.forward.x₁.ξ
+  yξ = mesh.cell_center_metrics.forward.x₂.ξ
+  zξ = mesh.cell_center_metrics.forward.x₂.ξ
+  xη = mesh.cell_center_metrics.forward.x₁.η
+  yη = mesh.cell_center_metrics.forward.x₂.η
+  zη = mesh.cell_center_metrics.forward.x₂.η
+  xζ = mesh.cell_center_metrics.forward.x₁.η
+  yζ = mesh.cell_center_metrics.forward.x₂.η
+  zζ = mesh.cell_center_metrics.forward.x₂.η
 
-# Get the grid metrics
-# @inline function metrics(m::CurvilinearGrid3D, (i, j, k)::NTuple{3,Real}, t::Real=0)
-#   _jacobian_matrix = checkeps(
-#     m._jacobian_matrix_func(i - m.nhalo, j - m.nhalo, k - m.nhalo, t)
-#   )
-#   J = det(_jacobian_matrix)
-
-#   inv_jacobian_matrix = inv(_jacobian_matrix)
-
-#   ξx = inv_jacobian_matrix[1, 1]
-#   ξy = inv_jacobian_matrix[1, 2]
-#   ξz = inv_jacobian_matrix[1, 3]
-#   ηx = inv_jacobian_matrix[2, 1]
-#   ηy = inv_jacobian_matrix[2, 2]
-#   ηz = inv_jacobian_matrix[2, 3]
-#   ζx = inv_jacobian_matrix[3, 1]
-#   ζy = inv_jacobian_matrix[3, 2]
-#   ζz = inv_jacobian_matrix[3, 3]
-
-#   vx, vy, vz = grid_velocities(m, (i, j, k), t)
-#   ξt = -(vx * ξx + vy * ξy + vz * ξz)
-#   ηt = -(vx * ηx + vy * ηy + vz * ηz)
-#   ζt = -(vx * ζx + vy * ζy + vz * ζz)
-
-#   ξ = Metric3D(ξx, ξy, ξz, ξt)
-#   η = Metric3D(ηx, ηy, ηz, ηt)
-#   ζ = Metric3D(ζx, ζy, ζz, ζt)
-
-#   return (; ξ, η, ζ, J)
-# end
-
-# ------------------------------------------------------------------
-# Conservative Grid Metrics; e.g. ξ̂x = ξx * J
-# ------------------------------------------------------------------
-
-# @inline function conservative_metrics(
-#   m::CurvilinearGrid3D, (i, j, k)::NTuple{3,Real}, t::Real=0
-# )
-#   _jacobian_matrix = checkeps(
-#     m._jacobian_matrix_func(i - m.nhalo, j - m.nhalo, k - m.nhalo, t)
-#   )
-#   J = det(_jacobian_matrix)
-
-#   inv_jacobian_matrix = inv(_jacobian_matrix)
-
-#   ξx = inv_jacobian_matrix[1, 1]
-#   ξy = inv_jacobian_matrix[1, 2]
-#   ξz = inv_jacobian_matrix[1, 3]
-#   ηx = inv_jacobian_matrix[2, 1]
-#   ηy = inv_jacobian_matrix[2, 2]
[…]
