+++ test/unit/test_2d_axisymmetric.jl
@@ -6,7 +6,7 @@
   nr, nz = (8, 10)
   snap_to_axis = true
   symmetry_axis = :x # rotate about the pole axis
-  mesh = AxisymmetricRectlinearGrid(
+  mesh = axisymmetric_rectlinear_grid(
     (r0, z0), (r1, z1), (nr, nz), :meg6_symmetric, snap_to_axis, symmetry_axis
   )
 
