### 2024-12-16 cbafb52
refactor the metric layout
