+++ test/unit/test_2d.jl
@@ -1,5 +1,5 @@
 @testset "2D Rectangular Mesh" begin
-  include("common.jl")
+  # include("common.jl")
 
   ni, nj = (40, 80)
   x0, x1 = (0, 2)
@@ -16,47 +16,47 @@
   @test mesh.domain_limits.node == (ilo=6, ihi=46, jlo=6, jhi=86)
   @test mesh.domain_limits.cell == (ilo=6, ihi=45, jlo=6, jhi=85)
 
-  @test all(mesh.cell_center_metrics.forward.J[domain] .≈ 0.00125)
-  @test all(mesh.cell_center_metrics.forward.x₁.ξ[domain] .≈ 0.05)
-  @test all(mesh.cell_center_metrics.forward.x₂.ξ[domain] .≈ 0.0)
-  @test all(mesh.cell_center_metrics.forward.x₁.η[domain] .≈ 0.0)
-  @test all(mesh.cell_center_metrics.forward.x₂.η[domain] .≈ 0.025)
+  @test all(mesh.cell_center_metrics.J[domain] .≈ 0.00125)
+  @test all(mesh.cell_center_metrics.x₁.ξ[domain] .≈ 0.05)
+  @test all(mesh.cell_center_metrics.x₂.ξ[domain] .≈ 0.0)
+  @test all(mesh.cell_center_metrics.x₁.η[domain] .≈ 0.0)
+  @test all(mesh.cell_center_metrics.x₂.η[domain] .≈ 0.025)
 
-  @test all(mesh.cell_center_metrics.inverse.ξ.x₁[domain] .≈ 20.0)
-  @test all(mesh.cell_center_metrics.inverse.ξ.x₂[domain] .≈ 0.0)
-  @test all(mesh.cell_center_metrics.inverse.η.x₁[domain] .≈ 0.0)
-  @test all(mesh.cell_center_metrics.inverse.η.x₂[domain] .≈ 40.0)
-  @test all(mesh.cell_center_metrics.inverse.ξ.t[domain] .≈ 0.0)
-  @test all(mesh.cell_center_metrics.inverse.η.t[domain] .≈ 0.0)
+  @test all(mesh.cell_center_metrics.ξ.x₁[domain] .≈ 20.0)
+  @test all(mesh.cell_center_metrics.ξ.x₂[domain] .≈ 0.0)
+  @test all(mesh.cell_center_metrics.η.x₁[domain] .≈ 0.0)
+  @test all(mesh.cell_center_metrics.η.x₂[domain] .≈ 40.0)
+  @test all(mesh.cell_center_metrics.ξ.t[domain] .≈ 0.0)
+  @test all(mesh.cell_center_metrics.η.t[domain] .≈ 0.0)
 
-  @test all(mesh.cell_center_metrics.inverse_normalized.ξ̂.x₁[domain] .≈ 0.025)
-  @test all(mesh.cell_center_metrics.inverse_normalized.ξ̂.x₂[domain] .≈ 0.0)
-  @test all(mesh.cell_center_metrics.inverse_normalized.η̂.x₁[domain] .≈ 0.0)
-  @test all(mesh.cell_center_metrics.inverse_normalized.η̂.x₂[domain] .≈ 0.05)
-  @test all(mesh.cell_center_metrics.inverse_normalized.ξ̂.t[domain] .≈ 0.0)
[…]
