+++ src/discretization_schemes/metrics.jl
@@ -208,15 +208,13 @@ function conservative_metric!(
 end
 
 """
-    conservative_metrics!(scheme::DiscretizationScheme, x::AbstractArray{T,3}, y::AbstractArray{T,3}, z::AbstractArray{T,3}, forward_metrics, inverse_metrics) where {T}
+    conservative_metrics!(scheme::DiscretizationScheme, metrics, x::AbstractArray{T,3}, y::AbstractArray{T,3}, z::AbstractArray{T,3}, domain) where {T}
 
 Compute all the inverse metrics ∂ξ̂ᵢ/∂xᵢ using the conservative scheme from Thomas P, Lombard C. AIAA J 1979;17(10):1030–7 (https://doi.org/10.2514/3.61273) 
 """
 function conservative_metrics!(
   scheme::DiscretizationScheme,
-  normalized_inverse_metrics,
-  inverse_metrics,
-  forward_metrics,
+  metrics,
   x::AbstractArray{T,3},
   y::AbstractArray{T,3},
   z::AbstractArray{T,3},
@@ -233,41 +231,41 @@ function conservative_metrics!(
   end
 
   @views begin
-    ξ̂x = normalized_inverse_metrics.ξ̂.x₁
-    η̂x = normalized_inverse_metrics.η̂.x₁
-    ζ̂x = normalized_inverse_metrics.ζ̂.x₁
+    ξ̂x = metrics.ξ̂.x₁
+    η̂x = metrics.η̂.x₁
+    ζ̂x = metrics.ζ̂.x₁
 
-    ξ̂y = normalized_inverse_metrics.ξ̂.x₂
-    η̂y = normalized_inverse_metrics.η̂.x₂
-    ζ̂y = normalized_inverse_metrics.ζ̂.x₂
+    ξ̂y = metrics.ξ̂.x₂
+    η̂y = metrics.η̂.x₂
+    ζ̂y = metrics.ζ̂.x₂
 
-    ξ̂z = normalized_inverse_metrics.ξ̂.x₃
-    η̂z = normalized_inverse_metrics.η̂.x₃
-    ζ̂z = normalized_inverse_metrics.ζ̂.x₃
+    ξ̂z = metrics.ξ̂.x₃
+    η̂z = metrics.η̂.x₃
+    ζ̂z = metrics.ζ̂.x₃
 
-    ξx = inverse_metrics.ξ.x₁
-    ηx = inverse_metrics.η.x₁
-    ζx = inverse_metrics.ζ.x₁
+    ξx = metrics.ξ.x₁
+    ηx = metrics.η.x₁
+    ζx = metrics.ζ.x₁
 
-    ξy = inverse_metrics.ξ.x₂
-    ηy = inverse_metrics.η.x₂
-    ζy = inverse_metrics.ζ.x₂
+    ξy = metrics.ξ.x₂
+    ηy = metrics.η.x₂
+    ζy = metrics.ζ.x₂
 
-    ξz = inverse_metrics.ξ.x₃
-    ηz = inverse_metrics.η.x₃
-    ζz = inverse_metrics.ζ.x₃
+    ξz = metrics.ξ.x₃
+    ηz = metrics.η.x₃
+    ζz = metrics.ζ.x₃
 
-    x_ξ = forward_metrics.x₁.ξ
-    x_η = forward_metrics.x₁.η
-    x_ζ = forward_metrics.x₁.ζ
+    x_ξ = metrics.x₁.ξ
+    x_η = metrics.x₁.η
+    x_ζ = metrics.x₁.ζ
 
-    y_ξ = forward_metrics.x₂.ξ
-    y_η = forward_metrics.x₂.η
-    y_ζ = forward_metrics.x₂.ζ
[…]
