+++ Project.toml
@@ -1,7 +1,7 @@
 name = "CurvilinearGrids"
 uuid = "58b19789-dd19-4aa2-839c-c763a7b10301"
 authors = ["Sam Miller <smil@lle.rochester.edu> and contributors"]
-version = "0.6.5"
+version = "0.7.0"
 
 [deps]
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
