### 2024-12-16 f44c9cd
struct arrays compat fix
