+++ Project.toml
@@ -32,7 +32,7 @@ MappedArrays = "0.4"
 OffsetArrays = "1"
 Polyester = "0.7"
 StaticArrays = "1"
-StructArrays = ">= 0.6"
+StructArrays = "0.3, 0.4, 0.5, 0.6, 0.7"
 UnPack = "1"
 WriteVTK = "1"
 julia = "1.7"
