### 2024-12-16 9c2e519
update the paper to reflect the most recent API
