### 2024-12-16 smillerc/CurvilinearGrids.jl#24 @smillerc
@DanielVandH I've made the changes you requested. Check out the latest changes on master (new release) in the manuscript. Thanks!
