### 2024-12-16 smillerc/CurvilinearGrids.jl#24 @DanielVandH
Thanks @smillerc. Can you point out what the changes were for the first two points? i.e.

  1. Maybe to make it clearer where the statement of need ends, an extra section heading for the Examples could be useful
  2. The statement of need does make it clear now who and what the package is for, but it's still not as clear how it relates to existing software. You mention the state of the Julia ecosystem but are there similar things that could be referenced from other languages to compare the work with? By compare I don't necessarily mean whether one is better than the other but just in terms of what's implemented potentially (having this in a language like Julia is definitely a good bonus over other languages e.g. Python to begin with!). This would be especially useful for readers who might not know what Julia is.

I'm looking at https://github.com/openjournals/joss-reviews/issues/7508#issuecomment-2546349148 but don't see any change in e.g. the statement of need. Thanks.
