+++ paper/paper.md
@@ -25,8 +25,11 @@ Finite-difference discretizations of partial differential equations are widespre
 
 `CurvilinearGrids.jl` is a Julia package designed to facilitate curvilinear coordinate transformations and the generation of metrics for arbitrary 1D, 2D, and 3D geometry. The API is designed for simple grid construction with minimal user effort; the backend handles the complicated book-keeping associated with metric computation and ensures strict adherance to geometric conservation laws when required. While Julia packages exist to facilitate standard finite-differences, there are few, if any, that provide conservative metrics for curvilinear transformations.
 
+Grid metric computation is typically direcetly embedded into a particular simulation framework, rather than provided as a standalone library, since the metric and PDE discretizations must be consistent. Third-party grid generation sofware exists to provide complex multi-block structured grids for multi-physics codes but metrics are not provided -- metric computation must be tailored to the discretization of the simulation library. `CurvilinearGrids.jl` is unique in that it can be imported into existing Julia simulation libraries and immediately provide the ability to use curvilinear geometry (as long as the governing equations are appropriately modified). New discretization schemes, like a 6th-order central finite-difference or 5th order upwind scheme, can be added in a straightforward manner to expand the capabilities of `CurvilinearGrids.jl`.
 
-Users provide `CurvilinearGrids.jl` with discrete coordinate points defined in real-space $(x,y,z)$, or functional forms, e.g `x(i,j), y(i,j)`. These coordinates are then used to compute the transformation (via Jacobian matrices) to a uniform computational coordinate space $(\xi,\eta,\zeta)$. A common example of this is to use a body-fit or conformal grid and transform it so that it becomes a uniform grid in $(\xi,\eta,\zeta)$. Standard finite-difference stencils can be used on the uniform transformed grid. \autoref{fig:grid} shows an example of a curvilinear grid defined in real-space and it's representation in computational space. 
+# Example Use
+
+Users provide `CurvilinearGrids.jl` with discrete coordinate points defined in real-space $(x,y,z)$. These coordinates are then used to compute the transformation to a uniform computational coordinate space $(\xi,\eta,\zeta)$. A common example of this is to use a body-fit or conformal grid and transform it so that it becomes a uniform grid in $(\xi,\eta,\zeta)$. Standard finite-difference stencils can be used on the uniform transformed grid. \autoref{fig:grid} shows an example of a curvilinear grid defined in real-space and it's representation in computational space. 
 
 ![Curvilinear grid transformation.\label{fig:grid}](mesh.png)
 
