### 2024-12-16 #7508 @DanielVandH
Thanks for making my changes @smillerc from https://github.com/smillerc/CurvilinearGrids.jl/issues/24. I've finished with my review and am happy with the state of the paper. I think it will make a good contribution to this space. I recommend acceptance.
