### 2024-12-17 e5f7299
add citation to Julia
