+++ paper/paper.md
@@ -23,7 +23,7 @@ Finite-difference discretizations of partial differential equations are widespre
 
 # Statement of Need
 
-`CurvilinearGrids.jl` is a Julia package designed to facilitate curvilinear coordinate transformations and the generation of metrics for arbitrary 1D, 2D, and 3D geometry. The API is designed for simple grid construction with minimal user effort; the backend handles the complicated book-keeping associated with metric computation and ensures strict adherance to geometric conservation laws when required. While Julia packages exist to facilitate standard finite-differences, there are few, if any, that provide conservative metrics for curvilinear transformations.
+`CurvilinearGrids.jl` is a Julia [@julia] package designed to facilitate curvilinear coordinate transformations and the generation of metrics for arbitrary 1D, 2D, and 3D geometry. The API is designed for simple grid construction with minimal user effort; the backend handles the complicated book-keeping associated with metric computation and ensures strict adherance to geometric conservation laws when required. While Julia packages exist to facilitate standard finite-differences, there are few, if any, that provide conservative metrics for curvilinear transformations.
 
 Grid metric computation is typically directly embedded into a particular simulation framework, rather than provided as a standalone library, since the metric and PDE discretizations must be consistent. Third-party grid generation sofware exists to provide complex multi-block structured grids for multi-physics codes but metrics are not provided -- metric computation must be tailored to the discretization of the simulation library. `CurvilinearGrids.jl` is unique in that it can be imported into existing Julia simulation libraries and immediately provide the ability to use curvilinear geometry (as long as the governing equations are appropriately modified). New discretization schemes, like a 6th-order central finite-difference or 5th order upwind scheme, can be added in a straightforward manner to expand the capabilities of `CurvilinearGrids.jl`.
 
