### 2024-12-17 a6402d7
fix rendering of equation
