### 2024-12-17 c3b9923
fix typo in manuscript
