### 2024-12-17 #7508 @jbytecode
@smillerc - The manuscript now looks better. Please create a tagged release in the software repository, e.g. `v1.2.3` and create a Zenodo archive using this version tag. Please tell me the version and the archive DOI so we can relate the permanent archive to the submission. The metadata in the archive should match those in the paper, for example the author list and their DOIs should match exactly. Thank you in advance.
