### 2024-12-17 3a9da93
cleanup; add metric getter functions
