+++ test/unit/test_2d.jl
@@ -1,4 +1,5 @@
-@testset "2D Rectangular Mesh" begin
+@testset "2D Rectangular Mesh"
+begin
   # include("common.jl")
 
   ni, nj = (40, 80)
