### 2024-12-17 a2dee0c
Add zenodo doi badge to readme
