### 2024-12-17 #7508 @smillerc
Here's the DOI
`10.5281/zenodo.14510381` at https://zenodo.org/records/14510381

The release version is `v0.7.1`
https://github.com/smillerc/CurvilinearGrids.jl/releases/tag/v0.7.1
