### 2024-12-17 4fed709
Update references.bib
