+++ paper/references.bib
@@ -1,5 +1,5 @@
 @article{thomas1979,
-  title        = {Geometric {{Conservation Law}} and {{Its Application}} to {{Flow Computations}} on {{Moving Grids}}},
+  title        = {Geometric Conservation Law and Its Application to Flow Computations on Moving Grids},
   author       = {Thomas, P. D. and Lombard, C. K.},
   date         = {1979-10},
   journaltitle = {AIAA Journal},
@@ -15,7 +15,7 @@
 }
 
 @article{visbal2002,
-  title        = {On the {{Use}} of {{Higher-Order Finite-Difference Schemes}} on {{Curvilinear}} and {{Deforming Meshes}}},
+  title        = {On the Use of Higher-Order Finite-Difference Schemes on Curvilinear and Deforming Meshes},
   author       = {Visbal, Miguel R and Gaitonde, Datta V},
   date         = {2002-09},
   journaltitle = {Journal of Computational Physics},
@@ -78,7 +78,7 @@
 }
 
 @book{huang2011,
-	title        = {Adaptive {{Moving Mesh Methods}}},
+	title        = {Adaptive Moving Mesh Methods},
 	author       = {Huang, Weizhang and Russell, Robert D.},
 	year         = 2011,
 	location     = {New York, NY},
