### 2024-12-17 #7508 @danielskatz
👋 @smillerc - I'm suggesting a few small changes in https://github.com/smillerc/CurvilinearGrids.jl/pull/27. Please merge this, or let me know what you disagree with.

Additionally, the section you have labelled as Funding is typically called Acknowledgments in JOSS papers. Can you change this as well? (It's not a requirement, if there's a strong reason you can't follow the JOSS style, but it would be nicer if you could.)
