### 2024-12-17 #7508 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
