### 2024-12-17 #7508 @danielskatz
Congratulations to @smillerc (Samuel Miller) on your publication!!

JOSS depends on volunteers, and we couldn't do this without you
